Extend block transformer metadata to int (#25)

+++ src/main/java/com/gtnewhorizons/postea/utility/ChunkFixerUtility.java
@@ -97,7 +97,7 @@ public class ChunkFixerUtility {
 
                     BlockConversionInfo output = getBlockReplacement(
                         blockId,
-                        (byte) metadata,
+                        metadata,
                         world,
                         x + chunkXPos + 1,
                         y + sectionY,
